+++ src/lib/__init__.py
@@ -11,7 +11,7 @@ from .expressions import ExpressionParser, DefaultAction
 from .apppaths import AppPaths
 from .spreadsheet import SpreadsheetGen
 from .bodefano import BodeFano
-from .stabcircle import StabilityCircle
+from .circles import StabilityCircle
 from .shortstr import shorten_string_list
 from .clipboard import Clipboard
 from .settings import SParamViewerAppSettings, Settings, PlotType, SmithNorm, TdResponse, YQuantity, PhaseProcessing, PhaseUnit, CsvSeparator, CursorSnap, ColorAssignment, Parameters, LogNegativeHandling, MainWindowLayout, LargeMatrixBehavior, GuiColorScheme